Renamed and moved test. Changed test param.
